Add vscode configuration to .gitignore
Since I am working using vscode and have some custom settings
for running tests, running and debugging go code I don't want
those settings to be commited to not pollute the repository
and not break any settings for other people.

To do that I add the .vscode folder to the .gitignore file.