Remove payments from invoice
In the test golb structure there is no payments data. Passing this
data into our KSeF parser results in an error. Since the payment data
is optional in KSeF we decided to comment out this part of the
code and address the issue after we add some tests to the existing code.