Add go-xsd-validate package
To make sure that the xml we are generating is valid we want
to validate it againts the provided xsd. This will also make
it simpler to find changes if the KSeF API changes.

To validate the xml we will be using go-xsd-validate package
so we add it into the project. We also use this opportunity
to do a little cleanup and remove all not used packages.