Remove unused optional fields
There are some optional fields in our current invoice that are
being left empty. This creates an invoice that is not valid
and will not be accepted by the KSeF API.

To make the invoice valid we remove ExchangeRate, FP, TP fields
until we are able to populate them with correct values.