Refactor and re-add payments to invoice
Earlier when generating an invoice without payments the generation
crashed. As a temporary fix we disabled the payments part of the invoice.
Now after adding some tests to the invoice we re-visit the payments.
We add checking that the required fields are set so that the conversion
doesn't crash.
We also re-work some of the code responsible for setting the paid and
partially paid markers based on the advance payments. Previously it
was checking the amount of advance payments and due dates to determine
if the invoice was paid. This was a mistake and didn't give correct
values. The new approach takes the calculated totals and sets the values
based on that.
We changed the name of PartialPayments to DueDates since this is a more
correct name for the value.
We also added tests for the payments logic