Remove footer from invoice
The footer part is optional, and at the moment we didn't provide it
with any data so it was always empty. We can re-add it later when
we have some data for the footer element, but at the moment we
are removing this element while we focus on supporting the basic
invoice.