Fix linter errors
The recent chnages to the linter exposed some tech debt that we
want to fix to be up to date with the best practices.
Here we fix all the linter errors that our new linters caught.
These are mostly missing comments for exported methods but it
also caught some unused parameters and we changed the visibility
of some methods since they are only used internally.