Change variable names to fit the standard
In Go, the trend is to always use just err instead of defining the use-case with a prefix.
As a rule, the error response is only ever used in the next few lines, so a short variable
name is preferred.

In addition for this example, bytes is the name of a standard package in Go, so people
tend to use data for []byte.