Add omitempty to fields that can be empty according to KSeF schema
Our current implementation did not mark every field that can be
omited as such. This could make it so empty fields would be still
present in the generated xml files which we don't want.

In the commit we mark optional fields with the omitempty keyword and
we also make some methods that are only used internally less visible
outside.