Authorize KSeF API session
We want to be able to post invoices into the KSeF API and get back
the UPO that proves that the invoice has been accepted by the system.

The first step is communicating with the API to get the interactive
session. To get the interactive session we need to have the context
in which we send the requests this context is the tax id of the company.
Next we need a valid token from ksef it can be generated from
ksef.mf.gov.pl/web for production or from ksef-test.mf.gov.pl/web for
test. The last thing is the public key provided from https://ksef.mf.gov.pl/ and
https://ksef-test.mf.gov.pl/.

To get the session token we first need to get the authorization challenge for
our context. Then we use the context, token and authorization challenge to
create a interactive session and get the session token used in other requests.