Send invoice to the KSeF API
We want to be able to post invoices into the KSeF API and get back
the UPO that proves that the invoice has been accepted by the system.

Here we prepare the request that posts the invoice to KSeF backend.
We also prepare a request that checks the invoice status.

With those requests we can post the invoice and check if it finished
processing.