Session management in KSeF API
We want to be able to post invoices into the KSeF API and get back
the UPO that proves that the invoice has been accepted by the system.

The UPO files are generated after the interactive session is terminated.
To get the files we first need to terminate the session then after it
finishes generating the UPO files we can get them from the session status request.

Here we the session termination and status requests used in the process.