Setup basic fields for credit note
We want to be able to correct invoices send into the KSeF API.
To correct the invoices we need to prepare a credit note.
There are a few differences between a regular VAT invoice and
a polish credit note. The first is the InvoiceType that changes
from VAT to KOR. The second is a CorrectionType it is used to
define when the correction should take legal effect, with values
1-3 meaning:
1 - Changes take effect in the date of the invoice. This is mostly used
if you made a mistake in the original invoice.
2 - Changes take effect in the date of the correction. This is used for
example when giving a discount or returning items.
3 - Changes take effect in a different date than any of the invoices or
the date is different for different elements.
There is also a optional field CorrectionReason where you can give a reason
for the correction.
The last element is data of the Invoice you are correcting. You need to give
the SequentialNumber and IssueDate of the original invoice. If the invoice
has a KSeF number you need to provide it and set the KsefNumberPresent flag
to 1. If the original invoice didn't have a KSeF number you need to set the
NoKsefNumberPresent flag to 1.