Make UnitDiscount field not visible when the value is zero
The original way we set the UnitDiscount field it was returning a 0.00
string which was causing the field to be visible in the invoice in
every case. This is not the way we want it to work. The field should
only be visible if the discount amount is different than zero.