Refactor setting debug on API Client to a OptFunc
In the earlier version we set the debug of the resty client to
true. This made the CLI command very verbose by default this
is ok when checking how the command works but may not be very
usefull in normal usage. We refactored this function into OptFunc
allowing the user to opt into using it.