Instead of saving the UPO use it to sign the envelope
As a requirement for PDF invoices in Poland we need to add a
QR code. The QR code is created using the KSeF id and a hash
of the invoice. We get both of those values when receiving the UPO.
We sign those values and the URL of the QR code into the head of the
envelope and return the document in the command line for further processing.