Handle international customer party id
In KSeF depending on customer location you need to fill in different
fields. For domestic customers we need to fill in the NIP number.
For customers inside the EU we need to fill in the UECode and UEVatNumber.
For non EU customers we need to fill in CountryCode and IDNumber and in
cases when there is no tax id number present we need to set the NoID flag.